+++ MyCalcLib/MyCalcLib/Interfaces/IOutputService.cs
@@ -2,6 +2,6 @@
 {
     public interface IOutputService
     {
-        void Print(int firstNumb, char operation, int secondNumb, int result);
+        void Print(double firstNumb, char operation, double secondNumb, double result);
     }
 }
